requirements.txt
coverage==3.7.1
Django==1.7.7
django-extensions==1.5.1
djangorestframework==3.1.0
gnureadline==6.3.3
ipython==3.0.0
mock==1.0.1
mysqlclient==1.3.6
nose==1.3.4
pytz==2014.10
requests==2.6.0
scorched==0.6
six==1.9.0
